requirements.txt
audioread==2.1.8
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
fuzzywuzzy==0.17.0
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
Google-Images-Search==1.0.1
httplib2==0.14.0
idna==2.8
Pillow==6.2.1
pyacoustid==1.1.7
pyasn1==0.4.7
pyasn1-modules==0.2.7
pyfiglet==0.8.post1
python-dotenv==0.10.3
python-Levenshtein==0.12.0
python-resize-image==1.1.19
requests==2.22.0
rsa==4.0
six==1.12.0
termcolor==1.1.0
uritemplate==3.0.0
urllib3==1.25.6
